requirements.txt
cloudscraper
mistralai
python-dotenv
